refactor: Match existing api service
